Move AttributeSplitSuggestion to a more adequate place
