+++ src/skmultiflow/trees/attribute_observer/nominal_attribute_class_observer.py
@@ -1,4 +1,4 @@
-from skmultiflow.trees.attribute_split_suggestion import AttributeSplitSuggestion
+from skmultiflow.trees.attribute_test import AttributeSplitSuggestion
 from skmultiflow.trees.attribute_test import NominalAttributeBinaryTest
 from skmultiflow.trees.attribute_test import NominalAttributeMultiwayTest
 from skmultiflow.trees.attribute_observer import AttributeObserver
