+++ src/skmultiflow/trees/attribute_observer/numeric_attribute_class_observer_binary_tree.py
@@ -1,6 +1,6 @@
 from skmultiflow.trees.attribute_observer import AttributeObserver
 from skmultiflow.trees.attribute_test import NumericAttributeBinaryTest
-from skmultiflow.trees.attribute_split_suggestion import AttributeSplitSuggestion
+from skmultiflow.trees.attribute_test import AttributeSplitSuggestion
 from collections import Counter
