+++ src/skmultiflow/trees/attribute_observer/numeric_attribute_regression_observer.py
@@ -1,7 +1,7 @@
 import numpy as np
 from skmultiflow.trees.attribute_observer import AttributeObserver
 from skmultiflow.trees.attribute_test import NumericAttributeBinaryTest
-from skmultiflow.trees.attribute_split_suggestion import AttributeSplitSuggestion
+from skmultiflow.trees.attribute_test import AttributeSplitSuggestion
 
 
 class NumericAttributeRegressionObserver(AttributeObserver):
