+++ src/skmultiflow/trees/attribute_test/__init__.py
@@ -1,7 +1,8 @@
+from .attribute_split_suggestion import AttributeSplitSuggestion
 from .instance_conditional_test import InstanceConditionalTest
 from .nominal_attribute_binary_test import NominalAttributeBinaryTest
 from .nominal_attribute_multiway_test import NominalAttributeMultiwayTest
 from .numeric_attribute_binary_test import NumericAttributeBinaryTest
 
-__all__ = ["InstanceConditionalTest", "NominalAttributeBinaryTest",
+__all__ = ["AttributeSplitSuggestion", "InstanceConditionalTest", "NominalAttributeBinaryTest",
            "NominalAttributeMultiwayTest", "NumericAttributeBinaryTest"]
