+++ src/skmultiflow/trees/nodes/efdtc_nodes.py
@@ -1,5 +1,5 @@
 import numpy as np
-from skmultiflow.trees.attribute_split_suggestion import AttributeSplitSuggestion
+from skmultiflow.trees.attribute_test import AttributeSplitSuggestion
 from skmultiflow.trees.attribute_observer import NominalAttributeClassObserver
 from skmultiflow.trees.attribute_observer import NumericAttributeClassObserverGaussian
 from skmultiflow.trees.nodes import SplitNode
